moved CheckLists from -Spec class to class with main
